some more explicit double->int conversions to silence the compiler
svn path=/trunk/koffice/; revision=167769

+++ kpresenter/kpresenter_dlg_config.cc
@@ -637,7 +637,7 @@ ConfigureDefaultDocPage::ConfigureDefaultDocPage(KPresenterView *_view, QVBox *b
     QLabel *fontTitle = new QLabel(i18n("Default font:"), fontContainer);
 
     font= new QFont( doc->defaultFont() );
-    font->setPointSize( KoTextZoomHandler::layoutUnitPtToPt( font->pointSize() ) );
+    font->setPointSize( (int)KoTextZoomHandler::layoutUnitPtToPt( font->pointSize() ) );
 
     QString labelName = font->family() + ' ' + QString::number(font->pointSize());
     fontName = new QLabel(labelName, fontContainer);
